+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - `<Tabe />` - lighter impl on unspecified `remember.key` prop
 
+### Fixed
+- `<Table />` - per-column initial hidden state on no-accessor columns
+
 ## [1.7.2] - 2020-08-07
 ### Fixed
 - `<Table />` - table can now update based on dinamically sortBy